requirements.txt
# ── Core web
fastapi==0.135.3
uvicorn[standard]==0.44.0
jinja2==3.1.6
python-multipart==0.0.24
aiofiles==25.1.0

# ── ORM / Banco
sqlalchemy==2.0.49
alembic==1.18.4

# ── Validação (pydantic-core 2.45+ tem wheel para Python 3.14)
pydantic[email]==2.12.5
pydantic-settings==2.13.1

# ── Auth (bcrypt direto — passlib 1.7.4 é incompatível com bcrypt>=4.0)
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
pyotp==2.9.0

# ── Workers
celery==5.6.3
redis==7.4.0

# ── Hardware / Sensores
pymodbus==3.12.1
paho-mqtt==2.1.0
httpx==0.28.1

# ── PDF (Fase 3)
weasyprint==68.1

# ── Testes
pytest==8.3.5
